requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# stable_baselines_agent/common/game_wrapper.py: 8,15
gym == 0.18.0

# stable_baselines_agent/defend_the_line/record_evaluate.py: 21
matplotlib == 3.3.4

# stable_baselines_agent/common/game_wrapper.py: 10
# stable_baselines_agent/common/i_reward_shaper.py: 8
# stable_baselines_agent/common/utils.py: 10
# stable_baselines_agent/defend_the_line/evaluate.py: 16
# stable_baselines_agent/defend_the_line/record_evaluate.py: 19
# stable_baselines_agent/simpler_basic/evaluate.py: 16
numpy == 1.20.1

# stable_baselines_agent/common/utils.py: 8
# stable_baselines_agent/defend_the_line/record_evaluate.py: 17
opencv_python == 4.5.1.48

# stable_baselines_agent/defend_the_line/evaluate.py: 20
# stable_baselines_agent/defend_the_line/record_evaluate.py: 24
# stable_baselines_agent/defend_the_line/train.py: 17,18
# stable_baselines_agent/simpler_basic/evaluate.py: 20
# stable_baselines_agent/simpler_basic/train.py: 17,18
stable_baselines == 2.10.1

# stable_baselines_agent/common/utils.py: 11
# stable_baselines_agent/defend_the_line/record_evaluate.py: 20
tensorflow_gpu == 1.15.0
# if you use CPU, then
# tensorflow == 1.15.0

# stable_baselines_agent/defend_the_line/evaluate.py: 15
# stable_baselines_agent/defend_the_line/record_evaluate.py: 16
# stable_baselines_agent/simpler_basic/evaluate.py: 15
tqdm == 4.57.0

# stable_baselines_agent/common/game_wrapper.py: 9
# stable_baselines_agent/defend_the_line/record_evaluate.py: 18
# stable_baselines_agent/scenarios_spec_test.py: 18
vizdoom == 1.1.8
